Article Box Style
Abstracted article box so it's style can be applied in other sass files by importing the partial articlebox.
Fixed a bug that was causing the program to crash when the setFeaturedArticle method was called on the server.
Made the articles list update each time the user deletes one from the database without a page refresh.